[NL] Avoid warning for type mismatch in comparison

+++ NumLib/Function/Interpolation.h
@@ -77,7 +77,8 @@ void shapeFunctionInterpolate(
     auto const num_nodal_dof = sizeof...(interpolated_values) + 1;
     auto const num_nodes = shape_matrix_N.size();
 
-    assert(num_nodes*num_nodal_dof == nodal_values.size());
+    assert(num_nodes * num_nodal_dof ==
+           static_cast<std::size_t>(nodal_values.size()));
     (void) num_nodal_dof; (void) num_nodes; // no warnings when not in debug build
 
     detail::shapeFunctionInterpolate<0>(nodal_values, shape_matrix_N, interpolated_value,
